Going back to using int arrays to store sample information, using short arrays was a bad idea

+++ src/audio/data/AudioData.java
@@ -37,7 +37,7 @@ public abstract class AudioData {
     
     public abstract boolean hasNextSample(int channel);
     
-    public abstract short[] getChunk(int samples, int channel);
+    public abstract int[] getChunk(int samples, int channel);
     
     public abstract int[] getSamples(double seconds, int sampleRate, int channel, double length);
